Flytte tilrettelegging-utils til riktig fil
